Pin user_side width and diagnose the remaining checksum gap
ScenarioBank.checksum hashes raw array bytes, and user_side came from
np.arange(n_users) % 2, whose dtype is int32 on Windows and int64 on Linux.
The frozen bank checksums were therefore unreproducible off Linux while the
channel data was identical. Pinning the width is a no-op on Linux, where the
default was already int64, and makes N=16 reproduce exactly.

The other four N still differ, and the cause is not a bug that can be fixed
without invalidating every frozen checksum in the repository. Exactly one
element out of 128,000 differs by 4.441e-16, one ULP, between numpy builds,
and a byte-exact digest turns that into a completely different hash. Rather
than rebase every recorded checksum on a quantised digest, the failure now
reports the largest elementwise gap and says whether it is rounding or a
real difference, so a rebuild on other hardware gets a diagnosis instead of
a bare mismatch.

The comparison had to capture the stored bank before generation, which
overwrites the file it would otherwise be compared against.

+++ src/star_ris_rsma/physics.py
@@ -20,7 +20,11 @@ def generate_channel(rng: np.random.Generator, n_users: int, n_ris: int) -> Chan
     h_direct = cn((n_users,), 0.35)
     g_br = cn((n_ris,), 1.0)
     h_ru = cn((n_users, n_ris), 0.75)
-    user_side = np.arange(n_users) % 2
+    # Pinned width. ScenarioBank.checksum hashes raw array bytes, and numpy's
+    # default integer is int32 on Windows and int64 on Linux, so leaving this
+    # to the platform made the frozen bank checksums unreproducible off Linux
+    # while the channel data itself was identical.
+    user_side = np.arange(n_users, dtype=np.int64) % 2
     return ChannelSample(h_direct=h_direct, g_br=g_br, h_ru=h_ru, user_side=user_side)
